# openapi.yml
openapi: 3.0.0
info:
  title: Stock Edu API
  version: 1.0.0
servers:
  - url: http://168.107.22.83:8080/api
    description: production (OCI VM direct, 8080 포트 노출)
  - url: http://localhost:8080/api
    description: local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshToken:
      type: apiKey
      in: header
      name: x-refresh-token
  schemas:
...
    Memo:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        template_type: { type: string }
        content: { type: object }
        created_at: { type: string, format: date-time }
security: [] # 전역 보안 강제 안 함; 엔드포인트별 지정
paths:
...
  /chapters:
    get:
      summary: 챕터 목록
      tags: [Chapters]
      security: [{ bearerAuth: [], refreshToken: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    chapter_id: { type: integer }
                    title: { type: string }
                    keyword: { type: string, nullable: true }
                    is_theory_completed: { type: integer }
                    is_quiz_completed: { type: integer }
...
  /memo/:
    get:
      summary: 유저 메모 전체 조회
      tags: [Memo]
      security: [{ bearerAuth: [], refreshToken: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  memos:
                    type: array
                    items: { $ref: "#/components/schemas/Memo" }
    put:
      summary: 메모 저장·갱신 (동일 엔드포인트)
      tags: [Memo]
      security: [{ bearerAuth: [], refreshToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  description: 신규 저장
                  properties:
                    template_type: { type: string, example: 일기 }
                    content: { type: object }
                - type: object
                  description: 기존 메모 갱신
                  properties:
                    id: { type: integer }
                    template_type: { type: string }
                    content: { type: object }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  memo: { $ref: "#/components/schemas/Memo" }
...